Add platform gate to sls run/emit
