+++ lib/plugins/emit/index.js
@@ -5,6 +5,7 @@ const BbPromise = require('bluebird');
 const fdk = require('@serverless/fdk');
 const path = require('path');
 const stdin = require('get-stdin');
+const getAuthToken = require('../../utils/getAuthToken');
 const userStats = require('../../utils/userStats');
 const chalk = require('chalk');
 
@@ -61,7 +62,10 @@ class Emit {
 
     this.hooks = {
       'emit:emit': () =>
-        BbPromise.bind(this).then(this.retrieveData).then(this.parseData).then(this.emitEvent),
+        BbPromise.bind(this)
+          .then(this.retrieveData)
+          .then(this.parseData)
+          .then(this.emitEvent),
     };
   }
 
@@ -116,6 +120,12 @@ class Emit {
   }
 
   emitEvent() {
+    const authToken = getAuthToken();
+    if (!authToken) {
+      throw new this.serverless.classes
+        .Error('Must be logged in to use this command. Please run `sls login`.');
+    }
+
     userStats.track('service_emitted');
     const url = this.options.url || 'http://localhost:4000';
     const eventGateway = fdk.eventGateway({
