+++ lib/plugins/platform/platform.js
@@ -10,9 +10,9 @@ const BbPromise = require('bluebird');
 const fsExtra = require('../../utils/fs/fse');
 const fetch = require('node-fetch');
 const chalk = require('chalk');
-const configUtils = require('../../utils/config');
 const functionInfoUtils = require('../../utils/functionInfoUtils');
 const createApolloClient = require('../../utils/createApolloClient');
+const getAuthToken = require('../../utils/getAuthToken');
 const selectServicePublish = require('../../utils/selectors/selectServicePublish');
 
 // NOTE Needed for apollo to work
@@ -91,17 +91,7 @@ class Platform {
   }
 
   getAuthToken() {
-    if (process.env.SERVERLESS_TOKEN) {
-      return process.env.SERVERLESS_TOKEN;
-    }
-
-    const userConfig = configUtils.getConfig();
-    const currentId = userConfig.userId;
-    const globalConfig = configUtils.getGlobalConfig();
-    if (globalConfig.users && globalConfig.users[currentId] && globalConfig.users[currentId].auth) {
-      return globalConfig.users[currentId].auth.id_token;
-    }
-    return null;
+    return getAuthToken();
   }
 
   publishService() {
