+++ lib/plugins/run/index.js
@@ -20,6 +20,8 @@ const registerFunctionsToEventGateway = require('./utils/registerFunctionsToEven
 const manageLocalEmulator = require('./utils/manageLocalEmulator');
 const manageEventGateway = require('./utils/manageEventGateway');
 
+const getAuthToken = require('../../utils/getAuthToken');
+
 class Run {
   constructor(serverless, options) {
     this.serverless = serverless;
@@ -64,6 +66,12 @@ class Run {
   run() {
     const EVENT_GATEWAY_VERSION = '0.5.14';
     const LOCAL_EMULATOR_VERSION = '0.1.18';
+
+    const authToken = getAuthToken();
+    if (!authToken) {
+      throw new this.serverless.classes
+        .Error('Must be logged in to use this command. Please run `sls login`.');
+    }
 
     let functionsDeployed = false;
     let functionsRegistered = false;
